add updateport repo & handler

+++ backend/src/repositories/portfolio.ts
@@ -51,44 +51,6 @@ class RepositoryPortfolio implements IRepositoryPortfolio {
 
   async updatePort(arg: {
     portId: number;
-    title?: string;
-    body?: string;
-    tag?: string[];
-    address?: string;
-    sub_district?: string;
-    district?: string;
-    province?: string;
-    postCode?: number;
-    updateAt: Date;
-    companyId: number;
-  }): Promise<IPort> {
-    const port = await this.db.portfolio.findFirst({
-      where: { portId: arg.portId, companyId: arg.companyId },
-    });
-
-    if (!port) {
-      return Promise.reject(`no such port ${arg.portId} not found`);
-    }
-
-    return await this.db.portfolio.update({
-      include: { postedBy: { select: { companyName: true } } },
-      where: { portId: arg.portId },
-      data: {
-        title: arg.title,
-        body: arg.body,
-        tag: arg.tag,
-        address: arg.address,
-        sub_district: arg.sub_district,
-        district: arg.district,
-        province: arg.province,
-        postCode: arg.postCode,
-        updateAt: arg.updateAt,
-      },
-    });
-  }
-
-  async updatePortImage(arg: {
-    portId: number;
     imageContents?: string[];
     imageContentUrls?: string[];
     title?: string;
@@ -110,8 +72,10 @@ class RepositoryPortfolio implements IRepositoryPortfolio {
       return Promise.reject(`no such port ${arg.portId} not found`);
     }
 
-    await deleteFile(port.imageContents);
-    await deleteFile(port.imageContentUrls);
+    if (arg.imageContentUrls && arg.imageContents) {
+      await deleteFile(port.imageContents);
+      await deleteFile(port.imageContentUrls);
+    }
 
     return await this.db.portfolio.update({
       include: { postedBy: { select: { companyName: true } } },
